chore: move endpoint to more relevant location
